+++ src/lib/pages/code-details/index.tsx
@@ -13,7 +13,7 @@ import { useCodeData } from "lib/model/code";
 import { useCodeStore } from "lib/providers/store";
 import { AmpEvent, AmpTrack } from "lib/services/amplitude";
 import { InstantiatePermission } from "lib/types";
-import { getFirstQueryParam, isCodeId } from "lib/utils";
+import { getCw2Info, getFirstQueryParam, isCodeId } from "lib/utils";
 
 import { CodeInfoSection } from "./components/CodeInfoSection";
 import { CTASection } from "./components/CTASection";
@@ -33,6 +33,8 @@ const CodeDetailsBody = observer(
     const { chainId, codeData, publicProject } = codeDataState;
 
     if (!codeData) return <InvalidCode />;
+
+    const cw2Info = getCw2Info(codeData.cw2Contract, codeData.cw2Version);
 
     return (
       <>
@@ -56,9 +58,21 @@ const CodeDetailsBody = observer(
             </Flex>
             <Flex gap={2}>
               <Text fontWeight={500} color="text.dark" variant="body2">
-                Code ID
+                Code ID:
               </Text>
               <ExplorerLink type="code_id" value={codeId.toString()} />
+            </Flex>
+            <Flex gap={2}>
+              <Text fontWeight={500} color="text.dark" variant="body2">
+                CW2 Info:
+              </Text>
+              <Text
+                color={cw2Info ? "text.main" : "text.disabled"}
+                variant="body2"
+                wordBreak="break-all"
+              >
+                {cw2Info ?? "N/A"}
+              </Text>
             </Flex>
           </Flex>
           <CTASection
@@ -70,6 +84,8 @@ const CodeDetailsBody = observer(
             }
             permissionAddresses={codeData.permissionAddresses ?? []}
             contractCount={undefined}
+            cw2Contract={undefined}
+            cw2Version={undefined}
           />
         </Flex>
         <Divider borderColor="pebble.700" my={12} />