+++ src/lib/pages/contract-details/components/tables/migration/index.tsx
@@ -66,10 +66,10 @@ export const MigrationTable = ({
     );
 
   const templateColumns =
-    "90px minmax(300px, 1fr) repeat(2, max(150px)) max(232px) max(180px)";
+    "90px minmax(300px, 1fr) minmax(220px, 1fr) repeat(2, max(150px)) max(232px) max(180px)";
 
   return (
-    <TableContainer overflow="visible">
+    <TableContainer>
       <MigrationHeader templateColumns={templateColumns} />
       {migrationHistories.map((history, idx) => (
         <MigrationRow